fix: la-134 split type from index and update relate import
