+++ test/__test__/builders/userBuilder.ts
@@ -1,7 +1,7 @@
+import { DEFAULT_LOCALE, LocaleType, ROLE, RoleType } from '@src/types/constants';
 import type { Document } from 'mongoose';
 import mongoose from 'mongoose';
 
-import { DEFAULT_LOCALE, type LocaleType, ROLE, RoleType } from '../../../src/config';
 import UserModel, { type User, USER_STATUS } from '../../../src/models/user';
 
 export interface UserDocument extends Document, User {
